ajout de la fonction shoot_bigm

+++ shoots.h
@@ -10,6 +10,6 @@ void set_lbm(uint8_t id, bool val);
 bool get_lbm(uint8_t id);
 void shoot_myspace(void);
 
-void set_shoot(uint8_t coo, uint8_t val);
+void set_shoot(uint8_t coo, float val);
 uint8_t get_shoot(uint8_t coo);
 #endif /* SHOOTS_H_ */
